+++ src/main/java/com/mposhatov/util/Calculator.java
@@ -4,12 +4,12 @@ import org.apache.commons.math3.random.RandomDataGenerator;
 
 public class Calculator {
 
-    public static long calculatePercentageOf(long percent, long value) {
+    public static int calculatePercentageOf(int percent, int value) {
         return percent * value / 100;
     }
 
-    public static long generateNumberFromTo(long from, long to) {
-        return from == to ? to : new RandomDataGenerator().nextLong(from, to);
+    public static int generateNumberFromTo(int from, int to) {
+        return from == to ? to : new RandomDataGenerator().nextInt(from, to);
     }
 
 }
